AllTheBacteria/scripts/requirements.txt
osfclient
aiohttp>=3.9
PyYAML>=6.0
pyarrow
orjson
duckdb